requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
SQLAlchemy==2.0.28
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.2.0
scikit-learn==1.3.2
pandas==2.3.2
fpdf2==2.7.2
python-multipart==0.0.6
aiofiles==23.1.0
